Refactor PLC into AbstractPLC with two concrete implementations and implement RadioTelescopeControllerTest

+++ ControlRoomApplication/ControlRoomApplication/Entities/Orientation/Orientation.cs
@@ -8,7 +8,7 @@ namespace ControlRoomApplication.Entities
     [Serializable]
     public class Orientation
     {
-        public Orientation(long azimuth, long elevation)
+        public Orientation(double azimuth, double elevation)
         {
             Azimuth = azimuth;
             Elevation = elevation;
